@@ -1,59 +1,66 @@
 {
   "name": "react-boilerplate",
   "version": "1.0.0",
-  "description": "Boilerplate for a React project using Redux.",
+  "description": "Boilerplate for an isomorphic React project.",
   "author": "Jonathan Rubins",
   "scripts": {
-    "cs": "jscs app config server",
     "lint": "eslint app config server"
   },
   "pre-commit": [
-    "cs",
     "lint"
   ],
+  "engines": {
+    "node": "4.5.0",
+    "npm": "2.15.9"
+  },
   "repository": {
     "type": "git",
     "url": "git@github.com:jrubins/react-boilerplate.git"
   },
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
-    "bootstrap-sass": "3.3.6",
-    "classnames": "2.2.3",
-    "express": "4.13.4",
-    "font-awesome": "4.5.0",
-    "jade": "1.11.0",
-    "jquery": "2.2.1",
-    "keymirror": "0.1.1",
-    "lodash": "4.6.0",
-    "mobile-detect": "1.3.1",
-    "moment": "2.11.2",
-    "react": "0.14.7",
-    "react-addons-css-transition-group": "0.14.7",
-    "react-dom": "0.14.7",
-    "react-redux": "4.4.1",
-    "react-router": "2.0.0",
-    "redux": "3.3.1",
-    "redux-thunk": "2.0.1",
-    "scroll-behavior": "0.4.0"
+    "bootstrap-sass": "3.3.7",
+    "classnames": "2.2.5",
+    "express": "4.14.0",
+    "font-awesome": "4.6.3",
+    "helmet": "2.1.3",
+    "history": "1.17.0",
+    "isomorphic-fetch": "2.2.1",
+    "lodash": "4.15.0",
+    "moment": "2.15.0",
+    "pm2": "2.0.12",
+    "pug": "2.0.0-beta6",
+    "react": "15.3.1",
+    "react-dom": "15.3.1",
+    "react-redux": "4.4.5",
+    "react-router": "2.8.1",
+    "react-router-scroll": "0.3.2",
+    "redux": "3.6.0",
+    "redux-thunk": "2.1.0",
+    "serve-favicon": "2.3.0",
+    "winston": "2.2.0"
   },
   "devDependencies": {
-    "babel-jscs": "3.0.0-beta1",
-    "babel-plugin-lodash": "2.3.0",
-    "babel-preset-es2015": "6.6.0",
-    "babel-preset-react": "6.5.0",
-    "babel-register": "6.6.5",
-    "babelify": "7.2.0",
-    "browserify": "13.0.0",
-    "connect-livereload": "0.5.4",
-    "eslint": "2.8.0",
-    "eslint-plugin-react": "4.1.0",
+    "babel-plugin-lodash": "3.2.8",
+    "babel-plugin-transform-es2015-destructuring": "6.9.0",
+    "babel-plugin-transform-object-rest-spread": "6.8.0",
+    "babel-preset-es2015": "6.14.0",
+    "babel-preset-react": "6.11.1",
+    "babel-register": "6.14.0",
+    "babelify": "7.3.0",
+    "browserify": "13.1.0",
+    "connect-livereload": "0.6.0",
+    "dotenv": "2.0.0",
+    "eslint": "3.5.0",
+    "eslint-plugin-react": "6.2.1",
     "gulp": "3.9.1",
-    "gulp-live-server": "0.0.28",
-    "gulp-sass": "2.2.0",
+    "gulp-live-server": "0.0.30",
+    "gulp-sass": "2.3.2",
     "gulp-util": "3.0.7",
-    "jscs": "3.0.3",
-    "pre-commit": "1.1.2",
-    "react-addons-perf": "0.14.7",
+    "loose-envify": "1.2.0",
+    "pre-commit": "1.1.3",
+    "react-addons-perf": "15.3.1",
+    "through2": "2.0.1",
     "vinyl-source-stream": "1.1.0",
     "watchify": "3.7.0"
   }
