@@ -1,5 +1,3 @@
-'use strict';
+export const SERVER_START_KEYWORD = 'Server Started!';
 
-module.exports = {
-    SERVER_START_KEYWORD: 'Server Started!'
-};
+export const ENV_DEV = 'development';
