@@ -1,21 +1,22 @@
-'use strict';
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { Router, useRouterHistory } from 'react-router';
-import createBrowserHistory from 'history/lib/createBrowserHistory';
-import useScroll from 'scroll-behavior/lib/useStandardScroll';
+import { Router, browserHistory, applyRouterMiddleware } from 'react-router';
+import useScroll from 'react-router-scroll/lib/useScroll';
 
-import configureStore from './store/configureStore';
-import routes from './routes';
+import { setLogger } from 'utils/logs';
+import configureStore from 'store/configureStore';
+import routes from 'routes';
 
-// This app history uses the Browser history and the standard scroll behavior to immitate
-// browser behavior with the scroll position when a page transition happens.
-const appHistory = useScroll(useRouterHistory(createBrowserHistory))();
+// Set our logger to be the browser console.
+setLogger(console);
 
 ReactDOM.render(
-    <Provider store={configureStore(window.__INITIAL_STATE__)}>
-        <Router history={appHistory} routes={routes} />
-    </Provider>
+  <Provider store={configureStore(window.__INITIAL_STATE__)}>
+    <Router
+      history={browserHistory}
+      routes={routes}
+      render={applyRouterMiddleware(useScroll())}
+    />
+  </Provider>
 , document.getElementById('app'));
